[Test] setex INTMAX ttl case.

+++ ssdb-1.9.2/src/client/SSDB_client.h
@@ -105,7 +105,7 @@ public:
 	/**
 	 * Set the value of the key, with a time to live.
 	 */
-	virtual Status setx(const std::string &key, const std::string &val, int ttl) = 0;
+	virtual Status setx(const std::string &key, const std::string &val, int64_t ttl) = 0;
 	virtual Status del(const std::string &key) = 0;
 	virtual Status incr(const std::string &key, int64_t incrby, int64_t *ret) = 0;
 	virtual Status decr(const std::string &key, int64_t incrby, int64_t *ret) = 0;
