+++ ssdb-1.9.2/src/client/SSDB_impl.h
@@ -42,7 +42,7 @@ public:
 	virtual Status setbit(const std::string &key, int64_t bitoffset, int on);
 	virtual Status getbit(const std::string &key, int64_t bitoffset, int64_t* ret);
 	virtual Status getset(const std::string &key, const std::string &val, std::string *getVal);
-	virtual Status setx(const std::string &key, const std::string &val, int ttl);
+	virtual Status setx(const std::string &key, const std::string &val, int64_t ttl);
 	virtual Status del(const std::string &key);
 	virtual Status incr(const std::string &key, int64_t incrby, int64_t *ret);
 	virtual Status decr(const std::string &key, int64_t incrby, int64_t *ret);
